refactor: move Ref and RefMut to separate files
