+++ src/meta/raft-store/src/sm_v002/leveled_store/mod.rs
@@ -15,6 +15,8 @@
 pub mod level_data;
 pub mod leveled_map;
 pub mod map_api;
+pub mod ref_;
+pub mod ref_mut;
 pub mod static_leveled_map;
 pub mod sys_data;
 pub mod sys_data_api;
