+++ src/meta/raft-store/src/sm_v002/sm_v002.rs
@@ -252,8 +252,6 @@ impl SMV002 {
     /// It does not check expiration of the returned entry.
     pub async fn get_kv(&self, key: &str) -> Option<SeqV> {
         let got = MapApiRO::<String>::get(&self.levels.leveled_ref(), key).await;
-        // let got = self.levels.leveled_ref().get(key).await;
-        // let got = MapApiRO::<String>::get(&self.levels, key).await;
         Into::<Option<SeqV>>::into(got)
     }
